Final prefix changes

+++ util/ext.py
@@ -9,10 +9,10 @@ class Util:
         self.bot = bot
         self.uptime = None
     
-    async def throw_error(self, context: commands.Context, text: str, view: discord.ui.View = None, defer: bool = True, bold: bool = True):
+    async def throw_error(self, i: discord.Interaction, text: str, view: discord.ui.View = None, defer: bool = True, bold: bool = True, ephemeral: bool = True):
         try:
-            m = await context.send(content=f'**{text}** <:bloboohcry:1011458104782758009>' if bold else text + ' <:bloboohcry:1011458104782758009>', view=view)
-            if defer:
+            m = await i.response.send_message(content=f'**{text}** <:bloboohcry:1011458104782758009>' if bold else text + ' <:bloboohcry:1011458104782758009>', view=view, ephemeral=ephemeral)
+            if defer and not ephemeral:
                 await sleep(7)
                 await m.delete()
             else:
@@ -20,10 +20,10 @@ class Util:
         except:
             pass
     
-    async def throw_fine(self, context: commands.Context, text: str, view: discord.ui.View = None, defer: bool = True, bold: bool = True):
+    async def throw_fine(self, i: discord.Interaction, text: str, view: discord.ui.View = None, defer: bool = True, bold: bool = True, ephemeral: bool = False):
         try:
-            m = await context.send(content=f'**{text}** <:blobheart:1011458084239056977>' if bold else text + ' <:blobheart:1011458084239056977>', view=view)
-            if defer:
+            m = await i.response.send_message(content=f'**{text}** <:blobheart:1011458084239056977>' if bold else text + ' <:blobheart:1011458084239056977>', view=view, ephemeral=ephemeral)
+            if defer and not ephemeral:
                 await sleep(7)
                 await m.delete()
             else:
@@ -31,15 +31,22 @@ class Util:
         except:
             pass
     
-    async def request(self, *, url: str, extract: Literal['json', 'read', 'text'] = 'json'):
+    async def request(self, *, url: str, params: dict = {}, extract: Literal['json', 'read', 'text'] = 'json', as_dict=False):
         try:
             async with aiohttp.ClientSession() as session:
-                async with session.get(url=url) as res:
+                async with session.get(url=url, params=params) as res:
                     if extract.lower() == 'json':
-                        return await res.json()
+                        return Response(await res.json(), res.status, res.content_type) if as_dict else await res.json()
                     elif extract.lower() == 'read':
-                        return await res.read()
+                        return Response(await res.read(), res.status, res.content_type) if as_dict else await res.read()
                     elif extract.lower() == 'text':
-                        return await res.text()
-        except:
+                        return Response(await res.text(), res.status, res.content_type) if as_dict else await res.text()
+        except Exception as err:
+            print(err)
             return None
+
+class Response:
+    def __init__(self, data, status, contettype) -> None:
+        self.data = data or None
+        self.status = status or None
+        self.content_type = contettype or None
